+++ matrices/dense_matrix.hpp
@@ -29,15 +29,15 @@ DenseMatrix<T>::DenseMatrix(const shared_ptr<BaseMatrix<T>> rhs)
 }
 
 template <typename T>
-DenseMatrix<T>::DenseMatrix(unsigned long m, unsigned long n)
+DenseMatrix<T>::DenseMatrix(uint32_t m, uint32_t n)
 {
   this->m_num_rows = m;
   this->m_num_columns = n;
   this->m_vectors = new MathVector<T>[this->m_num_rows];
-  for(unsigned long i = 0; i < m; i++)
+  for(uint32_t i = 0; i < m; i++)
   {
     this->m_vectors[i] = MathVector<T>(n);
-    for(unsigned long j = 0; j < this->m_vectors[i].capacity(); j++)
+    for(uint32_t j = 0; j < this->m_vectors[i].capacity(); j++)
       this->m_vectors[i].push(0);
   }
 }
@@ -55,7 +55,7 @@ MatrixType DenseMatrix<T>::type() const
 }
 
 template <typename T>
-MathVector<T>& DenseMatrix<T>::operator[](unsigned long index)
+MathVector<T>& DenseMatrix<T>::operator[](uint32_t index)
 {
   if(index < 0 || index >= this->m_num_rows)
     throw out_of_range("Out of Range: DenseMatrix");
@@ -64,7 +64,7 @@ MathVector<T>& DenseMatrix<T>::operator[](unsigned long index)
 }
 
 template <typename T>
-MathVector<T>& DenseMatrix<T>::operator[](unsigned long index) const
+MathVector<T>& DenseMatrix<T>::operator[](uint32_t index) const
 {
   if(index < 0 || index >= this->m_num_rows)
     throw out_of_range("Out of Range: DenseMatrix");
@@ -72,13 +72,13 @@ MathVector<T>& DenseMatrix<T>::operator[](unsigned long index) const
 }
 
 template <typename T>
-T DenseMatrix<T>::operator()(unsigned long row_index, unsigned long column_index) const
+T DenseMatrix<T>::operator()(uint32_t row_index, uint32_t column_index) const
 {
   return (*this)[row_index][column_index];
 }
 
 template <typename T>
-void DenseMatrix<T>::operator()(unsigned long row_index, unsigned long column_index, T element)
+void DenseMatrix<T>::operator()(uint32_t row_index, uint32_t column_index, T element)
 {
   (*this)[row_index][column_index] = element;
 }
@@ -87,8 +87,8 @@ template <typename T>
 shared_ptr<BaseMatrix<T>> DenseMatrix<T>::transpose() const
 {
   shared_ptr<BaseMatrix<T>> ret = make_shared<DenseMatrix<T>>(this->getNumColumns(), this->getNumRows());
-  for(unsigned long i = 0; i < ret->getNumRows(); i++)
-    for(unsigned long j = 0; j < ret->getNumColumns(); j++)
+  for(uint32_t i = 0; i < ret->getNumRows(); i++)
+    for(uint32_t j = 0; j < ret->getNumColumns(); j++)
       (*ret)[i][j] = (*this)(j, i);
   return ret;
 }
@@ -98,7 +98,7 @@ template <typename T>
 DenseMatrix<T> DenseMatrix<T>::operator*(double c) const
 {
   DenseMatrix ret(this->clone());
-  for(unsigned long i = 0; i < ret.getNumRows(); i++)
+  for(uint32_t i = 0; i < ret.getNumRows(); i++)
     ret[i] = c * ret[i];
   return ret;
 }
@@ -110,7 +110,7 @@ DenseMatrix<T> DenseMatrix<T>::operator+(const BaseMatrix<T>& rhs) const
     throw domain_error("Sizes not equal. + DenseMatrix");
 
   DenseMatrix ret(this->clone());
-  for(unsigned long i = 0; i < ret.getNumRows(); i++)
+  for(uint32_t i = 0; i < ret.getNumRows(); i++)
     ret[i] += rhs[i];
   return ret;
 }
@@ -123,7 +123,7 @@ DenseMatrix<T> DenseMatrix<T>::operator-(const BaseMatrix<T>& rhs) const
     throw domain_error("Sizes not equal. - DenseMatrix");
 
   DenseMatrix ret(this->clone());
-  for(unsigned long i = 0; i < ret.getNumRows(); i++)
+  for(uint32_t i = 0; i < ret.getNumRows(); i++)
     ret[i] -= rhs[i];
   return ret;
 }
@@ -135,9 +135,9 @@ DenseMatrix<T> DenseMatrix<T>::operator*(const BaseMatrix<T>& rhs) const
     throw domain_error("Matrix sizes not compatible: * DenseMatrix.");
 
   DenseMatrix ret(this->getNumRows(), rhs.getNumColumns());
-  for(unsigned long i = 0; i < ret.getNumRows(); i++)
-    for(unsigned long j = 0; j < ret.getNumColumns(); j++)
-      for(unsigned long k = 0; k < this->getNumColumns(); k++)
+  for(uint32_t i = 0; i < ret.getNumRows(); i++)
+    for(uint32_t j = 0; j < ret.getNumColumns(); j++)
+      for(uint32_t k = 0; k < this->getNumColumns(); k++)
         ret[i][j] += (*this)(i, k) * rhs(k, j);
 
   return ret;
@@ -147,7 +147,7 @@ template <typename T>
 shared_ptr<BaseMatrix<T>> DenseMatrix<T>::clone() const
 {
   shared_ptr<BaseMatrix<T>> ret = make_shared<DenseMatrix<T>>(this->m_num_rows, this->m_num_columns);
-  for (unsigned long i = 0; i < this->m_num_rows; i++)
+  for (uint32_t i = 0; i < this->m_num_rows; i++)
     (*ret)[i] = this->m_vectors[i];
   return ret;
 }
